Setup with NPM dependency
