--- .babelrc
@@ -1,24 +1,0 @@
-{
-  "presets": [
-    [
-      "@babel/env",
-      {
-        "targets": {
-          "node": "8",
-          "browsers": ">1%, not ie 11"
-        },
-        "exclude": [
-          "transform-regenerator"
-        ]
-      }
-    ]
-  ],
-  "plugins": [
-    ["@chainsafe/babel-plugin-inline-binary-import", {
-      "extensions": [
-        ".wasm"
-      ]
-    }]
-  ]
-}
-
