+++ .github/workflows/main.yml
@@ -1,0 +1,14 @@
+name: Main
+
+on: [pull_request, push]
+
+jobs:
+  build:
+    name: Build
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - name: Install dependencies
+        run: yarn
+      - name: Test
+        run: yarn test
