+++ .gitignore
@@ -1,3 +1,10 @@
-dist
 node_modules
 .idea
+.vscode
+yarn-error.log
+
+# Build target
+bls-eth-wasm/
+
+# Diff target
+dist/
