--- .gitmodules
@@ -1,6 +1,0 @@
-[submodule "bls"]
-	path = bls
-	url = https://github.com/herumi/bls.git
-[submodule "mcl"]
-	path = mcl
-	url = https://github.com/herumi/mcl.git
