--- .travis.yml
@@ -1,13 +1,0 @@
-language: node_js
-node_js:
-  - "8"
-script:
-  - mkdir emsdk
-  - cd emsdk
-  - curl -fsSL -O https://raw.githubusercontent.com/derhuerst/install-emsdk/master/install-emsdk.sh
-  - chmod a+x ./install-emsdk.sh
-  - ./install-emsdk.sh ./emsdk
-  - source ./emsdk/emsdk_env.sh
-  - cd ..
-  - yarn run build
-  - yarn run test
